requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pymongo==4.10.1
gunicorn==23.0.0
python-dotenv==1.0.1
pytest==8.3.3


